Implement adding, editing, and removing time lags to sequence relationships

+++ src/ifcopenshell-python/ifcopenshell/api/sequence/assign_lag_time.py
@@ -1,20 +1,23 @@
 class Usecase:
     def __init__(self, file, **settings):
         self.file = file
-        self.settings = {"rel_sequence": None, "lag_value": None, "duration_type": "NOTEDEFINED"}
+        self.settings = {"rel_sequence": None, "lag_value": None, "duration_type": "NOTDEFINED"}
         for key, value in settings.items():
             self.settings[key] = value
 
     def execute(self):
         lag_value = self.file.createIfcDuration(self.settings["lag_value"])
-        lag_time = self.file.create_entity("IfcLagTime",
+        lag_time = self.file.create_entity(
+            "IfcLagTime",
             **{
-            "DurationType": self.settings["duration_type"],
-            "LagValue": lag_value,
+                "DurationType": self.settings["duration_type"],
+                "LagValue": lag_value,
             }
         )
-        #Can an IfcLagTime entity be used by multiple sequence relationships?
         if self.settings["rel_sequence"].is_a("IfcRelSequence"):
-            if len(self.file.get_inverse(self.settings["rel_sequence"].TimeLag)) == 1:
+            if (
+                self.settings["rel_sequence"].TimeLag
+                and len(self.file.get_inverse(self.settings["rel_sequence"].TimeLag)) == 1
+            ):
                 self.file.remove(self.settings["rel_sequence"].TimeLag)
         self.settings["rel_sequence"].TimeLag = lag_time
